refactor(#1063): split ExecutorConfig into CoreExecutorConfig + InfraExecutorConfig
- CoreExecutorConfig: LogicExecutor, CheckedLogicExecutor, ExecutionPipeline,
  ExceptionTranslator, LoggingPolicy, taskExecutor (shared by all modules)
- InfraExecutorConfig: alertTaskExecutor, aiTaskExecutor, asyncExecutor,
  expectationComputeExecutor, operationalExecutor, backfillExecutor,
  VT executor shutdown (module-app only)
- ExecutorConfig: backward-compatible @Import(CoreExecutorConfig, InfraExecutorConfig)
- Lightweight modules (external-api, synchronizer, calculator) import CoreExecutorConfig only
- ExecutorConfigTest updated to use InfraExecutorConfig directly

Unblocks #1068 (VT ExecutorManager refactor).
